pyproject.toml
[project]
name = "cinematic-frame-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.2",
    "open-clip-torch>=2.30.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "polars>=1.22.0",
    "rich>=13.9.4",
    "torch>=2.6.0",
    "typer>=0.15.1",
]
